requirements.txt
greenlet==1.1.2
mysql-connector-python==8.0.29
numpy==1.22.4
pandas==1.4.2
protobuf==3.20.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
